Language models are storing counts now.
Former-commit-id: a98de6e1539975f9fd9c6c6a277e4d63bed7df6f

+++ src/main/scala/de/uni_potsdam/hpi/coheel/programs/CoheelProgram.scala
@@ -242,7 +242,7 @@ abstract class CoheelProgram[T]() extends ProgramDescription {
 				val model = lineSplit(1).split(' ').flatMap { entrySplit =>
 					val wordSplit = entrySplit.split('\0')
 					if (wordSplit.length == 2)
-						Some(wordSplit(0), wordSplit(1).toDouble)
+						Some(wordSplit(0), wordSplit(1).toInt)
 					else
 						None
 				}.toMap
